move jest tests to their own folder
