+++ src/Playwright tests/example.spec.ts
@@ -1,0 +1,7 @@
+import { test, expect } from "@playwright/test";
+
+test("basic test", async ({ page }) => {
+  await page.goto("https://playwright.dev/");
+  const title = page.locator(".navbar__inner .navbar__title");
+  await expect(title).toHaveText("Playwright");
+});
