+++ src/jest tests/App.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable testing-library/no-node-access */
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import App from "../../App";
+import App from "../App";
 import renderer from "react-test-renderer";
 
 test("Sanity Check", () => {
